Refactor code, add useFlag hook
