+++ src/hooks/useFlags/flagsmith/useFlagsmithFlags.test.tsx
@@ -3,7 +3,7 @@ import flagsmith from "flagsmith";
 import { FlagsmithProvider } from "flagsmith/react";
 import React from "react";
 
-import { useAllFlagsmithFlags, useFlagsmithFlags } from "./useFlagsmithFlags";
+import { useFlagsmithFlags } from "./useFlagsmithFlags";
 
 const wrapper = ({ children }) => (
   <FlagsmithProvider
@@ -82,55 +82,5 @@ describe("useFlagsmithFlags hook", () => {
 
     // Assert
     expect(result.current).toEqual({});
-  });
-});
-
-describe("useAllFlagsmithFlags hook", () => {
-  test("should return all flags from Flagsmith", () => {
-    // Arrange && Act
-    const features = {
-      flag_one: {
-        value: "1",
-        enabled: true,
-      },
-      flag_two: {
-        value: "2",
-        enabled: true,
-      },
-    };
-
-    jest.spyOn(flagsmith, "getAllFlags").mockImplementation(() => features);
-
-    const { result } = renderHook(() => useAllFlagsmithFlags(), {
-      wrapper,
-    });
-
-    // Assert
-    expect(result.current).toEqual([
-      {
-        name: "flagOne",
-        enabled: true,
-        value: "1",
-      },
-      {
-        name: "flagTwo",
-        enabled: true,
-        value: "2",
-      },
-    ]);
-  });
-
-  test("should return empty array when there is no flags", () => {
-    // Arrange && Act
-    const features = {};
-
-    jest.spyOn(flagsmith, "getAllFlags").mockImplementation(() => features);
-
-    const { result } = renderHook(() => useAllFlagsmithFlags(), {
-      wrapper,
-    });
-
-    // Assert
-    expect(result.current).toEqual([]);
   });
 });
